% Generated by roxygen2 (4.0.1): do not edit by hand
\name{onePeq}
\alias{onePeq}
\title{One parameter test of equivalence}
\usage{
onePeq(mu, sd = 1, n, epsilon, alpha = 0.05, null = 0)
}
\arguments{
\item{mu}{mean of the sample}

\item{sd}{true standard deviation}

\item{n}{sample size}

\item{epsilon}{margin of error}

\item{alpha}{significance level}

\item{null}{null hypothesis value}
}
\value{
1 for reject, 0 for FTR
}
\description{
Perform a one parameter test of equivalence for a single sample based on the
sample statistics
}
\details{
Consider X_i ~ (theta, sigma) with sigma known.  We test the hypothesis H:
abs(theta - null) >= epsilon against K: abs(theta - null) < epsilon.  From
Wellek 2010 4.1, the UMP test is to compare abs(mu) to n^(-1/2) * C where C
is the alphath quantile from a chisquare distribution with non centrality
parameter sqrt(n)*epsilon and one degree of freedom.

Wellek shows this for X_i ~ N(theta, sigma), but I think it should follow as
long as Xbar ~ N(theta, sigma).
}
\references{
Wellek 2010 Testing Statistical Hypotheses of Equivalence and Noninferiority
}

