% Generated by roxygen2 (4.0.1): do not edit by hand
\name{qeplot2}
\alias{qeplot2}
\title{Plot error bars for each chain}
\usage{
qeplot2(chains, prob, ...)
}
\arguments{
\item{chains}{chains}

\item{prob}{a single probability}

\item{...}{arguments to plot}
}
\value{
Nothing is returned
}
\description{
Plot error bars for each chain
}
\details{
Plot error bars for each chain
}

