% Generated by roxygen2 (4.0.1): do not edit by hand
\name{tost}
\alias{tost}
\title{Two one-sided test of equivalence for summary stats}
\usage{
tost(mu, sd, n, epsilon, alpha = 0.05, null = 0)
}
\arguments{
\item{mu}{description}

\item{se}{description}

\item{n}{description}

\item{delta}{description}

\item{alpha}{description}

\item{H0}{description}
}
\value{
1 for reject, 0 for FTR
}
\description{
My description
}
\details{
This is a
}

