% Generated by roxygen2 (4.0.1): do not edit by hand
\name{qed}
\alias{qed}
\title{Compute the quantile equivalence diagnostic}
\usage{
qed(X, prob, epsilon = 0.005, alpha = 0.05)
}
\arguments{
\item{X}{matrix of chains}

\item{prob}{null probability of observation below quantile quant}

\item{epsilon}{margin of error on quantile}

\item{alpha}{significance level}
}
\value{
returned
}
\description{
Perform the quantile equivalence diagnostic for a single chain
}
\details{
Argument X is a matrix where each column is a chain
}

