% Generated by roxygen2 (4.0.1): do not edit by hand
\name{qed}
\alias{qed}
\title{Compute the quantile equivalence diagnostic}
\usage{
qed(X, q, quant, epsilon = 0.01, alpha = 0.05)
}
\arguments{
\item{X}{chain}

\item{q}{quantile of interest}

\item{quant}{percentile of quantile}

\item{epsilon}{margin of error on quantile}

\item{alpha}{significance level}
}
\value{
returned
}
\description{
Perform the quantile equivalence diagnostic for a single chain
}
\details{
work in progress
}

