\name{qed_evolve}
\alias{qed_evolve}
\title{Display the evolution of the quantile of interest over the chain(s).}
\description{
    Display the evolution of the quantile of interest over the chain(s) at a
    specified step size.
}
\usage{fun(arg1, arg2, ...)}
\arguments{
    \item{arg}{description}
}
\details{
    Extension of description
}
\value{Return value, use item if list}
\references{Use \url{} of \href{}{} for web}
\note{
   Not necessary, Pie charts are very bad way of ....
}
\author{Name \email{myemail.email.com}}
\seealso{Point to code \code{\link{...}}}
\examples{
}
\keyword{zero or more keywords}
